Whack element registration with simulation really hard. Discover Species instead of registering
Remove registration of loggers, accumulators and meters.
Replace with registration of Data{Stuff,Stream,Thing}s containing the DataSource and what holds it.  Consumers are expected to get them and walk the streams (using SimulationDataAction -- what a name! -- if desired) looking for the Data elements they want to act on.