On DeviceBox, changed the EmptyBorder to a MatteBorder with the same color as it's parent.  This should not alter current implementations.
Added a method to set the border backround color.  The method will only
change the background color if the border is a MatteBorder.  Other border
types are possible for a DeviceBox and would ignore the request to set
the border background.

Used the ability to change DeviceBox border colors in crystal viewer.   Rather than set the text field background color for cell dimensions that are tied together, use the same color border around the DeviceBox for boxes whose values are tied.