Change layout manager of the Crystal Viewer panel the graphic display sits on to a BorderLayout.  This allows the graphic to be resized.
DisplayPhaseCanvas1D.setBounds : check size to make sure it is not <= 0.  If it is, do nothing.

DisplayPhase.setSize : check size to make sure it is not <= 0.  If it is, do nothing.

DisplayPhaseCanvasG3DSys .setBounds : check size to make sure it is not <= 0.  If it is, do nothing.

DisplayPhaseCanvasG3DSys .setBackground : Now also sets the background of the panel the graphic sits on.  Reduces the noticable flicker.

When resizing the Crystal Viewer, the graphic is flickering.  Need to get to the bottom of this flicker.