Bugzilla 194
Same change to TestHC2YukawaMC3D and TestYukawaMC3D :

The MCMove atom uses an AtomSourceRandomLeaf as its default atom source.  The applications were creating its own AtomSourceRandomLeaf instance and setting the MCMove atom source with this new instance.  The new instance was not setting a random number generator and was getting a null pointer exception when its use was attempted.

Removed setting a new atom source for MCMove and just use the default fixed the problem.