Bugzilla 209
Added some ctors for SimulationGraphic that will allow the interval between graphic area repaints to be set.  If the interval is not set (by calling a ctor that does not take the parameter) a default will be used.  The default will be set to the value used prior to this change.

The Osmosis application had its own redraw implementation.  Chaned to use as described above.