Bug 204
The drawing of atoms for a  DisplayPhaseSpin2D is a little different than for DisplayPhaseCanvas2D.  DisplayPhaseSpin2D extends DisplayPhaseCanvas2D but uses DisplayPhaseCanvas2D doPaint method.  So, the call to drawAtom is class instance dependent.

(Or could make changes to DisplayPhaseSpin2D.drawAtom similar to DisplayPhaseCanvas2D.drawAtom ,which relies on atoms position being set properly, and change DisplayPhaseSpin2D clients to set position of atoms correctly.  That was a lot more work!)