Bug 214
The DeviceSlider is having problems painting minor ticks.  Every time setTicks is called, new minor tick positions are computed, but the old minor tick positions do not seem to be erased.

Made change to "get around" this by not painting minor ticks unless the show minor values is explicitely set by the client.

So, any of the applications that use the DeviceSlider that were showing minor ticks and were getting lucky that the slider values they were using matched up with the defaults : well, your not going to get the minor ticks anymore until I do a little more work.