Bug 208
Removed the thread in Piston Cylinder that was controlling the painting of the graphic.  The painting now works off of an integrator action like all of the other applications that extend SimulationGraphic.

Had to move the "graphic Scale" slider off of the graphic display.  For now, it is on the control panel.  There is NO WAY I can leave it like this.  I need to figure out how to relabel the slider with text rather than with numbers.  Possibly the labels "min" and "max" will work.  Regardless, the current lablels of 10 and 100 cannot stay.