Bug 251
Made change to DevicePlotPoints so that the header names in the data
table can be changed at the device level.  The client should not have
to drill down into the table model to do this and then have to send a
table changed event.  It wasn't working at the table model anyway
because firing an event on the header row was changing the entire
table;s editor to the editor used for the header, which is read only.