Remove IAtom interface.  IAtomLeaf and IMolecule no longer have a common superclass.
Made changes to IteratorDirective to enable it to continue to hold both atoms and molecules.

Created an IBoxMoleculeEvent.  Removed IBoxAtomAddedEvent and IBoxAtomRemovedEvent.  In its place, added IBoxMoleculeAddedEvent and IBoxMoleculeRemovedEvent.  Added concrete classes for the interfaces.

Noticed that IntegratorHard.actionPerformed() was not calling randomizeMomenta() as the IBoxAtomAddedEvent was never sent.  Hopefully, fixed this problem.