Initial restructuring of simulation serialization framework.
SimulationModel is the access point for the serialized simulation. Wrapped simulation
objects are kept as a flat list, with simple consecutive index (id field in
Wrapper) used to reference each other. See EtomicaServer for main method demonstrating
use.

More reliance is made on Wrapper and Property implementations to handle different
cases.

Serialization produces a JSON file with most of the desired information.
Presently however it is missing some formatting elements so it is not
fully correct JSON (this should be simple to fix by someone who knows
what they're doing when writing this file).

Additional details need attention.