Refactor EventManager and uses
Change to concrete class with a type parameter
for event type, replace redundant concrete classes with instances
(they all did the same thing). Change internal structure from an
ad-hoc linked list to a CopyOnWriteArrayList, which eliminates the need
for synchronization.

Also removes DeviceTernarySelector since it's not used. If this class is
needed again it would require slight modifications to its event manager