Change IntegratorHard nullPotentialManager to map
This AtomTypeAgentManager has a source that just produces null and is
used for mapping AtomType to a potential. Since it doesn't produce
anything for new AtomTypes it doesn't need the SimulationListener
capabilities and can just be a hashmap. IntegratorHard doesn't need
to take a Simulation in the constructor now, but I'll make that a separate
commit so this one is less noisy.