Refactor AtomTypeAgentManager
Add type parameters and require a non null agent source and sim.
Internal map is a HashMap instead of the IndexMap that AtomLeafAgentManager
uses because there aren't enough AtomTypes for there to be any performance
gain.

Refactor all uses of the AgentIterator (in PotentialMasters) to use
native iteration over the map returned by getAgents.