Refactor BoxAgentManager
Allow the option of specifying just a Function to produce the agent,
since the release method in AgentSource is always empty. This way you
can use a lambda which could reduce the verbosity of AgentSources.
PotentialMasterNbr and family are a bit tangled in regards to
BoxAgentManagers so make sure this doesn't introduce any bugs there.