Change PotentialCalculationForceSum to take Vector agent manager
Instead of agent manager of Forcibles. This allows all the force vectors
to be packed together in one array which benchmarking has suggested
provides a speedup. Fixes some of the trivial uses of setAgentManager
but the rest do not yet compile. Also deletes the threaded package
(again) so it doesn't have to be updated.