Forcible to vectors (#99)
* Change PotentialCalculationForceSum to take Vector agent manager

Instead of agent manager of Forcibles. This allows all the force vectors
to be packed together in one array which benchmarking has suggested
provides a speedup. Fixes some of the trivial uses of setAgentManager
but the rest do not yet compile. Also deletes the threaded package
(again) so it doesn't have to be updated.

* Change remaining uses of Forcible agent managers given to PCForceSum

This is mostly changing everything that uses IntegratorVelocityVerlet.MyAgent
to use plain Vector agents (since MyAgent holds nothing except the
force vector). The only outliers which have Forcible agents that hold
other data are IntegratorDroplet and IntegratorVerlet, these now have
2 agent managers, one for the forces and one for the Agent.

Also fix null agent source in DisplayBoxG3DSys, now just has a source that
returns null.