Convert some MeterPotentialEnergy uses to pass Box in constructor
For now this just reduces verbosity and possibility of error, as this
isn't every use of MPE and the setBox can't be removed yet (mostly I'm
just unsure whether the MCMoves actually need a flexible box, in
which case this does too). Note this refactoring may have messed up
some formatting.