Use correct potential for a pair in NeighborIterator PotentialMasters
The old logic which loops over potentials is incorrect, now there is
only one potential per atom type combination. However, to avoid slowdown
from accessing both dimensions of the 2d potentials array from inside
the neighbor loop, introduce a second callback to the neighborIterator
interface which is called separately for the downlist neighbors. Since
in either callback we know which atom is the "other" atom we can lift
the first array dereference out of the inner loop.