Add initial implementation of SpeciesManager and Builder
Aims to replace species-collection functionality of Simulation.
Can then be passed to things that depend on the species data (e.g. Box,
PotentialComputes) without needing a whole Simulation. SpeciesManager
is immutable to enforce the fact that the species data must be fixed
throughout the sim lifetime. Species are added to the Builder and then
finalized into the SpeciesManager before they can be passed to anything.